add mappings square brackets to `r` for text object commands
Mapping similar to `b` for parens and `B` for curly brackets.
Square brackets on the keyboard is quite hard to reach, so a mapping on
the key near the middle of the keyboard will make this easier.

Maybe there is an elegant way to configure this? This will suffice for
now.